Move `@relation` directive to graphql-catalog plugin
