+++ plugins/graphql-catalog/src/catalog/catalog.graphql
@@ -1,3 +1,9 @@
+directive @relation(
+  name: String
+  nodeType: String
+  kind: String
+) on FIELD_DEFINITION
+
 scalar JSON
 scalar JSONObject
