+++ plugins/graphql-catalog/src/relation/relation.graphql
@@ -1,0 +1,5 @@
+directive @relation(
+  name: String
+  nodeType: String
+  kind: String
+) on FIELD_DEFINITION
