+++ plugins/graphql-catalog/src/relation/relation.ts
@@ -13,8 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { transformSchema } from '@backstage/plugin-graphql-common';
-import { printSchemaWithDirectives } from '@graphql-tools/utils';
-import { Catalog } from './catalog';
+import { createModule } from 'graphql-modules';
+import { relationSchema } from './schema';
+import { relationDirectiveMapper } from '../relationDirectiveMapper';
+import { createDirectiveMapperProvider } from '@backstage/plugin-graphql-common';
 
-export default printSchemaWithDirectives(transformSchema([Catalog]));
+/** @public */
+export const Relation = createModule({
+  id: 'relation',
+  typeDefs: relationSchema,
+  providers: [
+    createDirectiveMapperProvider('relation', relationDirectiveMapper),
+  ],
+});
